# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Medical Coding AI",
  "description": "AI-assisted medical coding: clinical documents -> evidence-backed codes -> validation -> human review.",
  "version": "1.0.0"
 },
 "paths": {
  "/api/v1/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login_api_v1_auth_login_post",
    "requestBody": {
     "content": {
      "application/x-www-form-urlencoded": {
       "schema": {
        "$ref": "#/components/schemas/Body_login_api_v1_auth_login_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/me": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Me",
    "operationId": "me_api_v1_auth_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_login_api_v1_auth_login_post": {
    "properties": {
     "grant_type": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^password$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Grant Type"
     },
     "username": {
      "type": "string",
      "title": "Username"
     },
     "password": {
      "type": "string",
      "format": "password",
      "title": "Password"
     },
     "scope": {
      "type": "string",
      "title": "Scope",
      "default": ""
     },
     "client_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client Id"
     },
     "client_secret": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "format": "password",
      "title": "Client Secret"
     }
    },
    "type": "object",
    "required": [
     "username",
     "password"
    ],
    "title": "Body_login_api_v1_auth_login_post"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Role": {
    "type": "string",
    "enum": [
     "admin",
     "coder",
     "auditor",
     "service"
    ],
    "title": "Role"
   },
...
   "TokenOut": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "token_type": {
      "type": "string",
      "title": "Token Type",
      "default": "bearer"
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In"
     },
     "user": {
      "$ref": "#/components/schemas/UserOut"
     }
    },
    "type": "object",
    "required": [
     "access_token",
     "expires_in",
     "user"
    ],
    "title": "TokenOut"
   },
...
   "UserOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "full_name": {
      "type": "string",
      "title": "Full Name"
     },
     "role": {
      "$ref": "#/components/schemas/Role"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "last_login_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Login At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "full_name",
     "role",
     "is_active",
     "created_at"
    ],
    "title": "UserOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   },
   "APIKeyHeader": {
    "type": "apiKey",
    "in": "header",
    "name": "X-API-Key"
   }
  }
 }
}